+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-- Added compatibility for Item Descriptions.
+Changes to portable repair with flint:
+- Repairing now consumes flint equal to the amount of repair materials instead of just one.
+- Flint can no longer be used to repair diamond and netherite items.
+  - This can be configured in the mod config. Adding repair material IDs to the flint material blacklist will prevent flint from repairing items which use those repair materials.